requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==19.7.1
idna==2.10
install==1.3.3
joblib==0.16.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.11.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
threadpoolctl==2.1.0
urllib3==1.25.9
virtualenv==20.0.28
whitenoise==5.2.0
